+++ docs/CHANGELOG.md
@@ -15,6 +15,10 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## TBD
+### Fixed
+- Prefixes not being imported or exported to Excel
+
 ## [0.92.2] - 17-09-24
 ### Added
 - Method in `InformationAnalysis` which returns class URI based on class entity